pyproject.toml
[tool.poetry]
name = "spark-static-stream-job"
version = "0.1.0"
repository = "https://github.com/leonidee/spark-static-stream-job"
description = ""
authors = [
    "Grishenkov Leonid <grishenkovleonid@gmail.com>"
]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
pyyaml = "^6.0"
kafka-python = "^2.0.2"
pandas = "^2.0.3"
findspark = "^2.0.1"
s3fs = "^2023.6.0"
pyarrow = "^12.0.1"

[tool.poetry.group.dev.dependencies]

[tool.poetry.group.testing.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
